+++ src/game/nlcore/action/actions/menuAction.ts
@@ -6,6 +6,7 @@ import type {CalledActionResult} from "@core/gameTypes";
 import {ContentNode} from "@core/action/tree/actionTree";
 import {TypedAction} from "@core/action/actions";
 import {Story} from "@core/elements/story";
+import {LogicAction} from "@core/action/logicAction";
 
 export class MenuAction<T extends typeof MenuActionTypes[keyof typeof MenuActionTypes] = typeof MenuActionTypes[keyof typeof MenuActionTypes]>
     extends TypedAction<MenuActionContentType, T, Menu> {
@@ -28,7 +29,7 @@ export class MenuAction<T extends typeof MenuActionTypes[keyof typeof MenuAction
         return awaitable;
     }
 
-    getFutureActions(story: Story) {
+    getFutureActions(story: Story): LogicAction.Actions[] {
         const menu = (this.contentNode as ContentNode<MenuActionContentType["menu:action"]>).getContent();
         return [...this.callee._getFutureActions(menu.choices), ...super.getFutureActions(story)];
     }