+++ src/game/player/elements/displayable/type.ts
@@ -1,24 +1,7 @@
-import {ElementProp, ITransition} from "@core/elements/transition/type";
-import React from "react";
 import {Transform} from "@core/elements/transform/transform";
-import {GameState} from "@player/gameState";
 import {EventDispatcher} from "@lib/util/data";
 import {Transition} from "@core/elements/transition/transition";
 
-/**@internal */
-export type DisplayableChildProps = {
-    transition: ITransition | null;
-    transform: Transform | null;
-    transformProps: ElementProp<Element, React.HTMLAttributes<Element>>;
-    transformRef: React.RefObject<HTMLDivElement | null>;
-    state: GameState;
-};
-/**@internal */
-export type DisplayableChildHandler = (props: Readonly<DisplayableChildProps>) => React.ReactElement;
-/**@internal */
-export type StatefulDisplayable = {
-    state: Record<any, any>;
-};
 /**@inetrnal */
 export type DisplayableAnimationEvents =
     | "event:displayable.applyTransform"